requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
defusedxml==0.6.0
Django==2.0
django-active-link==0.1.6
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-extensions==3.0.7
django-taggit==0.22.2
django-widget-tweaks==1.4.8
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.10